refactor(runtime): simplify readiness reporting
